[feat/#63]: 텍스트 크기 수정
